+++ src/backend/Cargo.toml
@@ -13,4 +13,4 @@ candid = "0.9.9"
 ic-cdk = "0.11.1"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1.0"
-ic-stable-structures = "0.5.6"
+ic-stable-structures = "0.5.5"
